updated uint32_t unix time to uint64_t  and fixed moods

+++ tempusludicus/Core/Inc/unixFunction.h
@@ -13,7 +13,7 @@ typedef struct {
     uint16_t second;
 } datetime_t;
 
-void RTC_HAL_convert_unix_to_datetime(const uint32_t seconds, datetime_t *datetime);
-void RTC_HAL_convert_datetime_to_unix(const datetime_t *datetime, uint32_t *seconds);
+void RTC_HAL_convert_unix_to_datetime(const uint64_t seconds, datetime_t *datetime);
+void RTC_HAL_convert_datetime_to_unix(const datetime_t *datetime, uint64_t *seconds);
 
 #endif
